added dashboard state management for widgets
